clean up and tuning
