--- annin_ar4_driver/config/controller_update_rate.yaml
@@ -1,3 +1,0 @@
-controller_manager:
-  ros__parameters:
-    update_rate: 125 # Hz
