+++ annin_ar4_driver/src/run_arduino_nano.cpp
@@ -5,7 +5,7 @@
 
 int main() {
   annin_ar4_driver::ArduinoNanoDriver driver;
-  bool success = driver.init("/dev/ttyUSB0", 9600);
+  bool success = driver.init("/dev/ttyUSB0", 112500);
   if (!success) {
     return -1;
   }
